1. A brand-new UI 2. Better navigation for captcha process and authentication process 3. Adjustment for deletion process to avoid accidental deletion 4. Abandoning mvc's default servlet handler: serving and saving files at appointed locations 5. Everything will be using docker: Nginx, Tomcat, MySQL, and Redis.

+++ ClientEnd/src/main/java/com/example/service/CourseService.java
@@ -21,15 +21,32 @@ public class CourseService {
         return courseDao.addCourse(name, desc, termId, inp, mr);
     }
 
-    public boolean deleteById(Integer id) {
-        return courseDao.deleteById(id);
+    public String deleteById(Integer id) {
+        // first check preconditions to delete:
+        // 1. no note is associated with this course
+        // 2. the course is deletable
+        if (courseDao.checkNotesExistForCourse(id) ){
+            return "couldn't delete course because there are notes associated with it";
+        }
+        if (!courseDao.isDeletable(id)) {
+            return "couldn't delete course because the course is not deletable";
+        }
+        courseDao.deleteById(id);
+        return "success";
     }
 
     public Course getById(Integer id) {
         return courseDao.getById(id);
     }
 
-    public boolean updateCourse(Integer id, String name, String desc, Integer termId, Boolean inp, Boolean mr) {
-        return courseDao.updateCourse(id, name, desc, termId, inp, mr);
+    public boolean updateCourse(Integer id, String name, String desc, Integer termId, Boolean inp, Boolean mr, Boolean deletable) {
+        return courseDao.updateCourse(id, name, desc, termId, inp, mr, deletable);
+    }
+
+    public List<Course> getCoursesByTermId(Integer termId) {
+        if (termId == null || termId == 0) {
+            return courseDao.getAllCourses();
+        }
+        return courseDao.getCoursesByTerm(termId);
     }
 }
